# public/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Polling Booth","description":"API of Polling Booth"},"schemes":["http"],"host":"localhost:3000","basePath":"/","securityDefinitions":{"authentication":{"type":"apiKey","name":"Authorization","in":"header"}},"tags":[{"name":"user"},{"name":"election"},{"name":"candidate"},{"name":"ballot"}],"paths":{"/users/verification-code":{"post":{"tags":["user"],"summary":"Fetchs verification code","parameters":[{"in":"formData","name":"email","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success' }"},"422":{"description":"Unprocessable entity"},"1001":{"description":"Too frequent requests","example":"{ code: 1001, message: 'too frequent requests' }"}}}},"/users/register":{"post":{"tags":["user"],"summary":"Register","parameters":[{"in":"formData","name":"email","schema":{"type":"string"},"required":true},{"in":"formData","name":"fullname","schema":{"type":"string"},"required":true},{"in":"formData","name":"password","schema":{"type":"string"},"required":true},{"in":"formData","name":"verificationCode","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { user: {}, token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoxMjMsImlhdCI6MTQ5MTQ3NTQyNCwiZXhwIjoxNDkxNDc1NDg0fQ.hYNC4qFAyhZClmPaLixfN137d41R2CFk1xPlfLK10JU' } }"},"422":{"description":"Unprocessable entity"},"1001":{"description":"Verification code is incorrect","example":"{ code: 1001, message: 'verification code is incorrect' }"},"1002":{"description":"Email address already in use","example":"{ code: 1002, message: 'email address already in use' }"}}}},"/users/login":{"post":{"tags":["user"],"summary":"Login","parameters":[{"in":"formData","name":"email","schema":{"type":"string"},"required":true},{"in":"formData","name":"password","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { user: {}, token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoxMjMsImlhdCI6MTQ5MTQ3NTQyNCwiZXhwIjoxNDkxNDc1NDg0fQ.hYNC4qFAyhZClmPaLixfN137d41R2CFk1xPlfLK10JU' } }"},"422":{"description":"Unprocessable entity"},"1001":{"description":"User does not exist","example":"{ code: 1001, message: 'user does not exist' }"},"1002":{"description":"Password is incorrect","example":"{ code: 1002, message: 'password is incorrect' }"}}}},"/elections":{"get":{"tags":["election"],"summary":"List elections","parameters":[{"in":"query","name":"cursor","schema":{"type":"integer"}},{"in":"query","name":"limit","schema":{"type":"integer"}}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { elections: [] } }"}}},"post":{"security":[{"authentication":[]}],"tags":["election"],"summary":"Create election","parameters":[{"in":"formData","name":"title","schema":{"type":"string"},"required":true},{"in":"formData","name":"description","schem":{"type":"string"},"required":true},{"in":"formData","name":"startTime","schema":{"type":"integer"},"required":true},{"in":"formData","name":"endTime","schema":{"type":"integer"},"required":true},{"in":"formData","name":"disabled","schema":{"type":"boolean"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { election: {} } }"}}}},"/elections/{id}":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"get":{"tags":["election"],"summary":"Show election","responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { election: {} } }"}}},"patch":{"security":[{"authentication":[]}],"tags":["election"],"summary":"Update election","parameters":[{"in":"formData","name":"title","schema":{"type":"string"},"required":true},{"in":"formData","name":"description","schem":{"type":"string"},"required":true},{"in":"formData","name":"startTime","schema":{"type":"integer"},"required":true},{"in":"formData","name":"endTime","schema":{"type":"integer"},"required":true},{"in":"formData","name":"disabled","schema":{"type":"boolean"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { election: {} } }"}}}},"/elections/{electionId}/candidates":{"parameters":[{"in":"path","name":"electionId","required":true,"schema":{"type":"integer"}}],"get":{"tags":["candidate"],"summary":"List candidates of election","parameters":[{"in":"query","name":"cursor","schema":{"type":"integer"}},{"in":"query","name":"limit","schema":{"type":"integer"}}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { candidates: [] } }"}}},"post":{"security":[{"authentication":[]}],"tags":["candidate"],"summary":"Create candidate","parameters":[{"in":"formData","name":"fullname","schema":{"type":"string"},"required":true},{"in":"formData","name":"avatar","schema":{"type":"string"},"required":true},{"in":"formData","name":"introduction","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { candidate: {} } }"}}}},"/elections/{electionId}/candidates/{id}":{"parameters":[{"in":"path","name":"electionId","required":true,"schema":{"type":"integer"}},{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"get":{"tags":["candidate"],"summary":"Show candidate","responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { candidate: {} } }"}}},"patch":{"security":[{"authentication":[]}],"tags":["candidate"],"summary":"Update candidate","parameters":[{"in":"formData","name":"fullname","schema":{"type":"string"},"required":true},{"in":"formData","name":"avatar","schema":{"type":"string"},"required":true},{"in":"formData","name":"introduction","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success', data: { candidate: {} } }"}}},"delete":{"security":[{"authentication":[]}],"tags":["candidate"],"summary":"Destroy candidate","responses":{"200":{"description":"Success","example":"{ code: 200, message: 'success' }"}}}}}}